Make SIGINT cleanly exit the emulator.
This hack is a patch I've run on snes9x for years ...  I run snes9x in
a custom setup where it gets SIGINT to terminate.  This would lead to
saveram files being in an inconsistent state.

Since saving the saveram is not signal safe, I introduce this hacky
bool which gets checked during the graphics loop.